Move S3 implementation to its own class
