+++ core/src/main/scala/quasar/destination/s3/S3Destination.scala
@@ -24,36 +24,16 @@ import quasar.api.resource.ResourcePath
 import quasar.connector.{MonadResourceErr, ResourceError}
 import quasar.contrib.pathy.AFile
 
-import cats.Applicative
-import cats.effect.{Concurrent, ConcurrentEffect, ContextShift, ExitCase}
+import cats.effect.Concurrent
 import cats.syntax.applicative._
 import cats.syntax.flatMap._
 import cats.syntax.functor._
 import eu.timepit.refined.auto._
-import fs2.Stream
-import monix.catnap.syntax._
 import pathy.Path
 import scalaz.NonEmptyList
-import software.amazon.awssdk.core.async.AsyncRequestBody
-import software.amazon.awssdk.services.s3.S3AsyncClient
-import software.amazon.awssdk.services.s3.model.{
-  AbortMultipartUploadRequest,
-  AbortMultipartUploadResponse,
-  CompleteMultipartUploadRequest,
-  CompleteMultipartUploadResponse,
-  CompletedMultipartUpload,
-  CompletedPart,
-  CreateMultipartUploadRequest,
-  CreateMultipartUploadResponse,
-  UploadPartRequest
-}
 
-class S3Destination[F[_]: ConcurrentEffect: ContextShift: MonadResourceErr](
-  client: S3AsyncClient,
-  config: S3Config,
-  partSize: Int) extends Destination[F] {
-  import S3Destination.{ensureAbsFile, push}
-
+class S3Destination[F[_]: Concurrent: MonadResourceErr](bucket: Bucket, upload: Upload[F])
+    extends Destination[F] {
   def destinationType: DestinationType = DestinationType("s3", 1L)
 
   def sinks: NonEmptyList[ResultSink[F]] =
@@ -61,19 +41,12 @@ class S3Destination[F[_]: ConcurrentEffect: ContextShift: MonadResourceErr](
 
   private def csvSink = ResultSink.csv[F](RenderConfig.Csv()) {
     case (path, _, bytes) =>
-      ensureAbsFile(path).flatMap(push(client, bytes, config, _, partSize))
+      for {
+        afile <- ensureAbsFile(path)
+        key = Path.posixCodec.printPath(nestResourcePath(afile)).drop(1)
+        _ <- upload.push(bytes, bucket, key)
+      } yield ()
   }
-}
-
-object S3Destination {
-  def apply[F[_]: ConcurrentEffect: ContextShift: MonadResourceErr](
-    client: S3AsyncClient,
-    config: S3Config,
-    partSize: Int): S3Destination[F] =
-    new S3Destination[F](client, config, partSize)
-
-  private def ensureAbsFile[F[_]: Applicative: MonadResourceErr](r: ResourcePath): F[AFile] =
-    r.fold(_.pure[F], MonadResourceErr[F].raiseError(ResourceError.notAResource(r)))
 
   private def nestResourcePath(file: AFile): AFile = {
     val withoutExtension = Path.fileName(Path.renameFile(file, _.dropExtension)).value
@@ -83,96 +56,12 @@ object S3Destination {
     parent </> Path.dir(withoutExtension) </> Path.file1(withExtension)
   }
 
-  private def push[F[_]: Concurrent](
-    client: S3AsyncClient,
-    bytes: Stream[F, Byte],
-    config: S3Config,
-    path: AFile,
-    partSize: Int): F[Unit] = {
-    val key = Path.posixCodec.printPath(nestResourcePath(path)).drop(1)
+  private def ensureAbsFile(r: ResourcePath): F[AFile] =
+    r.fold(_.pure[F], MonadResourceErr[F].raiseError(ResourceError.notAResource(r)))
+}
 
-    Concurrent[F].bracketCase(
-      startUpload(client, config.bucket, key))(createResponse =>
-      for {
-        parts <- uploadParts(client, bytes, createResponse.uploadId, partSize, config.bucket, key)
-        _ <- completeUpload(client, createResponse.uploadId, config.bucket, key, parts)
-      } yield ()) {
-      case (createResponse, ExitCase.Canceled | ExitCase.Error(_)) =>
-        abortUpload(client, createResponse.uploadId, config.bucket, key).void
-      case (_, ExitCase.Completed) =>
-        Concurrent[F].unit
-    }
-  }
-
-  private def startUpload[F[_]: Concurrent](client: S3AsyncClient, bucket: String, key: String)
-      : F[CreateMultipartUploadResponse] =
-    Concurrent[F].delay(client.createMultipartUpload(
-      CreateMultipartUploadRequest.builder.bucket(bucket).key(key).build)).futureLift
-
-  private def uploadParts[F[_]: Concurrent](
-    client: S3AsyncClient,
-    bytes: Stream[F, Byte],
-    uploadId: String,
-    minChunkSize: Int,
-    bucket: String,
-    key: String): F[List[CompletedPart]] =
-    (bytes.chunkMin(minChunkSize).zipWithIndex evalMap {
-      case (byteChunk, n) => {
-        // parts numbers must start at 1
-        val partNumber = Int.box(n.toInt + 1)
-
-        val uploadPartRequest =
-          UploadPartRequest.builder
-            .bucket(bucket)
-            .uploadId(uploadId)
-            .key(key)
-            .partNumber(partNumber)
-            .contentLength(Long.box(byteChunk.size.toLong))
-            .build
-
-        val uploadPartResponse =
-          Concurrent[F].delay(
-            client.uploadPart(
-              uploadPartRequest,
-              AsyncRequestBody.fromByteBuffer(byteChunk.toByteBuffer))).futureLift
-
-        uploadPartResponse.map(response =>
-          CompletedPart.builder.partNumber(partNumber).eTag(response.eTag).build)
-      }
-    }).compile.toList
-
-  private def completeUpload[F[_]: Concurrent](
-    client: S3AsyncClient,
-    uploadId: String,
-    bucket: String,
-    key: String,
-    parts: List[CompletedPart]): F[CompleteMultipartUploadResponse] = {
-    val multipartUpload =
-      CompletedMultipartUpload.builder.parts(parts :_*).build
-
-    val completeMultipartUploadRequest =
-      CompleteMultipartUploadRequest.builder
-        .bucket(bucket)
-        .key(key)
-        .uploadId(uploadId)
-        .multipartUpload(multipartUpload)
-        .build
-
-    Concurrent[F].delay(
-      client.completeMultipartUpload(completeMultipartUploadRequest)).futureLift
-  }
-
-  private def abortUpload[F[_]: Concurrent](
-    client: S3AsyncClient,
-    uploadId: String,
-    bucket: String,
-    key: String): F[AbortMultipartUploadResponse] =
-    Concurrent[F].delay(
-      client.abortMultipartUpload(
-        AbortMultipartUploadRequest
-          .builder
-          .uploadId(uploadId)
-          .bucket(bucket)
-          .key(key)
-          .build)).futureLift
+object S3Destination {
+  def apply[F[_]: Concurrent: MonadResourceErr](bucket: Bucket, upload: Upload[F])
+      : S3Destination[F] =
+    new S3Destination[F](bucket, upload)
 }
