+++ core/src/main/scala/quasar/destination/s3/Upload.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2014–2019 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.destination.s3
+
+import slamdata.Predef._
+
+import fs2.Stream
+
+trait Upload[F[_]] {
+  def push(bytes: Stream[F, Byte], bucket: Bucket, key: String): F[Unit]
+}
